% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTableauPipeline-package.R
\docType{package}
\name{GTableauPipeline-package}
\alias{GTableauPipeline}
\alias{GTableauPipeline-package}
\title{GTableauPipeline: Automated Tableau Dashboard Generation}
\description{
Automatically Generate Tableau Dashboard from Template Dashboard More Content Goes Here
}
\keyword{internal}
