% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipeline.R
\name{Template}
\alias{Template}
\title{Automated Tableau Visualization Main Pipeline}
\usage{
Template(
  abk_0,
  adsl_0 = NA,
  definition_0 = NA,
  save_path = getwd(),
  UID = "SUBJID",
  study_name = NA,
  special_name = NA,
  about = NA,
  URL = NA,
  distinct_threshold = 4,
  fold_change = NA,
  variable_numerical = NA,
  variable_categorical = NA,
  variable_keep = c("COUNTRY"),
  generate_correlation = FALSE,
  debug_mode = FALSE,
  save_RDS = FALSE,
  trash_can = FALSE,
  connection = "Live",
  shinny = F,
  reference_line = NA,
  message = "No Message"
)
}
\arguments{
\item{abk_0}{ABK Dataframe, or Path}

\item{adsl_0}{ADSL Dataframe, or Path}

\item{UID}{Unique IDs to Identify Subjects Shared Between ABK and ADSL}

\item{study_name}{Study Name, Optional}

\item{special_name}{A Special Name to Rename Dashboard, Optional}

\item{about}{Study Information, Optional}

\item{URL}{Website URL, 'http' Not Supported, Optional}

\item{reference_line}{Reference Line Default Value, Optional}

\item{message}{Creator Message}
}
\value{
List of Waste Variables
}
\description{
Automatically Generate Tableau Dashboard from Template Dashboard
}
\examples{
Template(abk_0, adsl_0, about)
}
