% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTableauPipeline-package.R
\docType{package}
\name{GTableauPipeline-package}
\alias{GTableauPipeline}
\alias{GTableauPipeline-package}
\title{GTableauPipeline: Automated Tableau Dashboard Generation Pipeline}
\description{
This pipeline provides an automatically Tableau dashboard generator with a user-friendly GUI
}
\keyword{internal}
